Updated the countdown layout to the new version. Now there is not a viewpager, there is only one layout for the countdown and recaudats
