+++ .idea/assetWizardSettings.xml
@@ -19,8 +19,8 @@
                                   <option name="values">
                                     <map>
                                       <entry key="assetType" value="IMAGE" />
-                                      <entry key="imageAsset" value="C:\Users\Albert Risco\Desktop\group_2.png" />
-                                      <entry key="outputName" value="ic_left_arrow_white" />
+                                      <entry key="imageAsset" value="C:\Users\Albert Risco\Desktop\ic_plus.png" />
+                                      <entry key="outputName" value="ic_plus" />
                                       <entry key="theme" value="HOLO_DARK" />
                                       <entry key="themeColor" value="ed3525" />
                                     </map>
