+++ app/src/main/res/layout/fragment_magic_line.xml
@@ -15,8 +15,6 @@
         android:orientation="vertical"
         >
 
-        <include layout="@layout/toolbar_magic_line_top" />
-
         <ScrollView
             android:layout_width="match_parent"
             android:layout_height="wrap_content">
@@ -26,7 +24,7 @@
                 android:layout_height="wrap_content"
                 android:orientation="vertical">
 
-                <include layout="@layout/view_pager_countdown_recaudats" />
+                <include layout="@layout/layout_countdown_recaudats_participants" />
 
                 <include layout="@layout/layout_news" />
