+++ app/src/main/res/layout/layout_countdown_recaudats_participants.xml
@@ -1,0 +1,23 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="match_parent"
+    android:layout_height="@dimen/countdown_parent_height"
+    android:orientation="vertical"
+    android:background="@drawable/countdown_background"
+    >
+
+    <include layout="@layout/layout_countdown" />
+
+    <ImageView
+        android:layout_width="match_parent"
+        android:layout_height="1dp"
+        android:background="@color/white"
+        />
+
+    <include layout="@layout/layout_recaudats_participants" />
+
+
+
+
+</LinearLayout>
