+++ app/src/main/res/layout/layout_recaudats_participants.xml
@@ -1,0 +1,99 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="match_parent"
+    android:layout_height="@dimen/height_recaudats"
+    android:orientation="horizontal"
+    android:weightSum="100"
+    >
+
+    <RelativeLayout
+        android:layout_width="0dp"
+        android:layout_height="match_parent"
+        android:paddingTop="6dp"
+        android:layout_weight="43.2"
+        >
+
+        <TextView
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="274.543€"
+            style="@style/recaudats_num"
+            android:id="@+id/recaudats_num"
+            />
+        <TextView
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="@string/recaudats_text"
+            style="@style/recaudats_text"
+            android:textAllCaps="true"
+            android:layout_below="@id/recaudats_num"
+            />
+
+    </RelativeLayout>
+    
+    <ImageView
+        android:layout_width="1dp"
+        android:layout_height="match_parent" 
+        android:background="@color/white"
+        />
+    
+    <RelativeLayout
+        android:layout_width="0dp"
+        android:layout_height="match_parent"
+        android:paddingTop="6dp"
+        android:layout_weight="38.4"
+        >
+
+        <TextView
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="38.543"
+            style="@style/recaudats_num"
+            android:id="@+id/participants_num"
+            />
+        <TextView
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content" 
+            android:text="@string/participants_text"
+            style="@style/recaudats_text"
+            android:textAllCaps="true"
+            android:layout_below="@id/participants_num"
+            
+            />
+
+    </RelativeLayout>
+
+    <ImageView
+        android:layout_width="1dp"
+        android:layout_height="match_parent"
+        android:background="@color/white"
+        />
+    
+    <RelativeLayout
+        android:layout_width="0dp"
+        android:layout_height="match_parent"
+        android:layout_weight="17.9"
+        >
+        
+        <ImageView
+            android:layout_width="25dp"
+            android:layout_height="25dp"
+            android:src="@drawable/ic_plus"
+            android:layout_marginTop="15dp"
+            android:layout_centerHorizontal="true"
+            android:id="@+id/ic_plus"
+            />
+        <TextView
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="@string/mes_info"
+            style="@style/recaudats_text"
+            android:layout_below="@id/ic_plus"
+            />
+
+
+    </RelativeLayout>
+    
+
+</LinearLayout>
