+++ app/src/main/res/layout/toolbar_magic_line_top.xml
@@ -14,7 +14,7 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:text="@string/principal_top_toolbar"
-            style="@style/TopToolbar"
+            style="@style/TopTitle"
             android:id="@+id/top_toolbar_text"
             />
         <ImageView
