+++ app/src/main/res/values/dimensions.xml
@@ -2,8 +2,10 @@
 <resources>
 
     <!--CountDown-->
-    <dimen name="countdown_timer_height">249dp</dimen>
+    <dimen name="countdown_parent_height">299dp</dimen>
+    <dimen name="countdown_timer_height">232dp</dimen>
     <dimen name="padding_right_countdown">249dp</dimen>
+    <dimen name="margin_top_grid">105dp</dimen>
 
     <!--Recaudats-->
     <dimen name="margin_top_recaudats_num">66dp</dimen>
@@ -11,10 +13,12 @@
     <dimen name="margin_top_btn_fer_donacio">17.5dp</dimen>
     <dimen name="btn_fer_donacio_width">135dp</dimen>
     <dimen name="btn_fer_donacio_height">38dp</dimen>
+    <dimen name="height_recaudats">64dp</dimen>
         <!-- View Pager Indicator-->
         <dimen name="margin_bottom_indicator">12dp</dimen>
 
-
+    <!--Participants-->
+    <dimen name="width_participants">144dp</dimen>
 
     <!--Top Toolbar-->
     <dimen name="principal_top_toolbar_height">46dp</dimen>
