+++ app/src/main/res/values/styles.xml
@@ -8,7 +8,7 @@
         <item name="colorAccent">@color/colorAccent</item>
     </style>
 
-    <style name="TopToolbar" parent="Theme.AppCompat.Light.NoActionBar">
+    <style name="TopTitle" parent="Theme.AppCompat.Light.NoActionBar">
         <!-- Customize your theme here. -->
         <item name="colorPrimary">@color/colorPrimary</item>
         <item name="colorPrimaryDark">@color/colorPrimaryDark</item>
@@ -16,7 +16,7 @@
         <item name="android:textColor">@color/white</item>
         <item name="android:textSize">24sp</item>
         <item name="android:fontFamily">@font/lato_blackitalic</item>
-        <item name="android:layout_centerInParent">true</item>
+        <item name="android:layout_centerHorizontal">true</item>
     </style>
 
 
@@ -31,15 +31,17 @@
         <item name="android:fontFamily">@font/lato_light</item>
     </style>
 
+
+
     <!--Princial Recaudats-->
     <style name="recaudats_num" parent="Theme.AppCompat.Light.NoActionBar">
         <item name="colorPrimary">@color/colorPrimary</item>
         <item name="colorPrimaryDark">@color/colorPrimaryDark</item>
         <item name="colorAccent">@color/colorAccent</item>
         <item name="android:textColor">@color/white</item>
-        <item name="android:textSize">43.2sp</item>
+        <item name="android:textSize">26sp</item>
         <item name="android:fontFamily">@font/lato_light</item>
-        <item name="android:textAlignment">center</item>
+        <item name="android:layout_centerHorizontal">true</item>
     </style>
 
     <style name="recaudats_text" parent="Theme.AppCompat.Light.NoActionBar">
@@ -47,10 +49,10 @@
         <item name="colorPrimaryDark">@color/colorPrimaryDark</item>
         <item name="colorAccent">@color/colorAccent</item>
         <item name="android:textColor">@color/white</item>
-        <item name="android:textSize">12.6sp</item>
+        <item name="android:textSize">12sp</item>
         <item name="android:fontFamily">@font/lato_regular</item>
-        <item name="android:textAlignment">center</item>
-        <item name="android:textAllCaps">true</item>
+        <item name="android:layout_centerHorizontal">true</item>
+
     </style>
 
     <style name="btn_recaudats_text" parent="Theme.AppCompat.Light.NoActionBar">
